requirements.txt
asgiref==3.7.2
Django==5.0.2
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
inflection==0.5.1
packaging==24.2
pillow==11.0.0
PyJWT==2.10.0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.4.4
typing_extensions==4.9.0
uritemplate==4.1.1
whitenoise==6.8.2
